add confirm before delete

+++ app/Http/Controllers/Admin/AttributeController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Size;
 use App\Color;
+use App\Stock;
 
 class AttributeController extends Controller
 {
@@ -33,12 +34,24 @@ class AttributeController extends Controller
 
     public function removeSize($id) {
         $size = Size::findOrFail($id);
+        $temp = Stock::where('size_id', $id)->get();
+        
+        if(count($temp)) {
+            return redirect()->back()->with(['alert' => 'Không thể xóa size do đang có sản phẩm sử dụng size này']);
+        }
+
         $size->delete();
         return redirect('/admin/product-attribute/sizes');
     }
 
     public function removeColor($id) {
         $color = Color::findOrFail($id);
+
+        $temp = Stock::where('color_id', $id)->get();
+        if(count($temp)) {
+            return redirect()->back()->with(['alert' => 'Không thể xóa màu sắc do đang có sản phẩm sử dụng màu sắc này']);
+        }
+
         $color->delete();
         return redirect('/admin/product-attribute/colors');
     }
